% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_sf}
\alias{acs_sf}
\alias{acs5_2013}
\alias{acs5_2014}
\alias{acs5_2015}
\alias{acs5_2016}
\alias{acs5_2017}
\title{Shapes and ACS estimates for Boone County, MO.}
\format{
\code{sf} objects.

An object of class \code{sf} (inherits from \code{data.frame}) with 87 rows and 9 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 87 rows and 9 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 85 rows and 9 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 87 rows and 9 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 87 rows and 9 columns.
}
\usage{
acs5_2013

acs5_2014

acs5_2015

acs5_2016

acs5_2017
}
\description{
An \code{sf} object with ACS estimates for:
\itemize{
\item Boone County, Missouri
\item Table B19013
\item Block group level geography
\item Years 2013 - 2017
}
}
\details{
This dataset was assembled using shapefiles from the \code{tigris} package
and ACS estimates from the American FactFinder website on 2/28/2019.
See \code{data-prep-aff.R} in the Columbia example code for details.
American FactFinder has since been deprecated, and similar data are
available at \url{http://data.census.gov}.
}
\keyword{datasets}
