% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cdist}
\alias{cdist}
\title{Calculate pairwise distances between two sets of points}
\usage{
cdist(X, Y)
}
\arguments{
\item{X}{Matrix of coordinates for first set of points}

\item{Y}{Matrix of coordinates for second set of points}
}
\value{
Distance matrix between X and Y points
}
\description{
Calculate pairwise distances between two sets of points
}
\keyword{internal}
