% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{columbia_neighbs}
\alias{columbia_neighbs}
\title{City of Columbia neighborhoods.}
\format{
An \code{sf} object with 4 features (neighborhoods).
}
\usage{
columbia_neighbs
}
\description{
An \code{sf} object containing the geometry of four neighborhoods in the
City of Columbia, Boone County, Missouri. Based on shapefiles provided by
the Office of Information Technology / GIS, City of Columbia, Missouri.
}
\keyword{datasets}
