% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{counts2z_block}
\alias{counts2z_block}
\title{Calculate z-scores for count blocks}
\usage{
counts2z_block(idx, Y, mu, th, rng_block, eps = 1e-06)
}
\arguments{
\item{idx}{Index of genes}

\item{Y}{Expression matrix}

\item{mu}{Mean parameter vector}

\item{th}{Theta parameter vector}

\item{rng_block}{Random number block}

\item{eps}{Small value to prevent zeros}
}
\value{
Matrix of z-scores
}
\description{
Calculate z-scores for count blocks
}
\keyword{internal}
