% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{detect_spatial_genes}
\alias{detect_spatial_genes}
\title{Detect spatial genes}
\usage{
detect_spatial_genes(expr, W, gene_names, moran_thr = 0.15)
}
\arguments{
\item{expr}{Expression matrix}

\item{W}{Spatial weights}

\item{gene_names}{Gene names}

\item{moran_thr}{Threshold for Moran's I}
}
\value{
Vector of spatial gene names
}
\description{
Detect spatial genes
}
\keyword{internal}
