% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_stcos.R
\name{DIC}
\alias{DIC}
\title{Deviance Information Criterion}
\usage{
DIC(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments.}
}
\value{
A numeric value of the DIC.
}
\description{
Generic function to calculate Deviance Information Criterion (DIC) for a
given model object.
}
