% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_batch_count}
\alias{generate_batch_count}
\title{Generate batch effects in counts}
\usage{
generate_batch_count(true_count, p0, celltype_proportion, seq_depth)
}
\arguments{
\item{true_count}{True count matrices}

\item{p0}{Dropout probability}

\item{celltype_proportion}{Cell type proportions}

\item{seq_depth}{Sequencing depth}
}
\value{
List of count matrices with batch effects
}
\description{
Generate batch effects in counts
}
