% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{generate_cell_prop}
\alias{generate_cell_prop}
\title{Generate cell proportion layers}
\usage{
generate_cell_prop(layer_coord, layer_param)
}
\arguments{
\item{layer_coord}{Coordinate matrix with layer assignments}

\item{layer_param}{Layer parameters matrix}
}
\value{
Matrix of cell type proportions
}
\description{
Generate cell proportion layers
}
