% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{generate_covariance}
\alias{generate_covariance}
\title{Generate covariance matrix for genes}
\usage{
generate_covariance(gene_relationship)
}
\arguments{
\item{gene_relationship}{Matrix indicating gene relationships}
}
\value{
List of covariance matrices per cell type
}
\description{
Generate covariance matrix for genes
}
