% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{generate_covariance_group}
\alias{generate_covariance_group}
\title{Generate group-based covariance matrix}
\usage{
generate_covariance_group(gene_relationship)
}
\arguments{
\item{gene_relationship}{Matrix indicating gene group relationships}
}
\value{
Array of covariance matrices
}
\description{
Generate group-based covariance matrix
}
