% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{generate_layer}
\alias{generate_layer}
\title{Generate layer coordinates}
\usage{
generate_layer(coord, xmin, xmax, ymin, ymax)
}
\arguments{
\item{coord}{Base coordinates}

\item{xmin}{Vector of minimum x values}

\item{xmax}{Vector of maximum x values}

\item{ymin}{Vector of minimum y values}

\item{ymax}{Vector of maximum y values}
}
\value{
Data frame with layer assignments
}
\description{
Generate layer coordinates
}
