% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{generate_pattern}
\alias{generate_pattern}
\title{Generate expression pattern}
\usage{
generate_pattern(
  coord,
  pattern,
  center_coord = NULL,
  center_param = NULL,
  background_param = NULL,
  hot_spot_size = NULL,
  streak_x = NULL,
  streak_x_param = NULL,
  streak_size = NULL,
  layer = NULL,
  layer_param = NULL,
  seed = 123
)
}
\arguments{
\item{coord}{Coordinates data.table}

\item{pattern}{Pattern type ("hot_spot", "streak", "layer")}

\item{center_coord}{Center coordinates for hotspot}

\item{center_param}{Center parameters}

\item{background_param}{Background parameters}

\item{hot_spot_size}{Size of hotspots}

\item{streak_x}{X positions for streaks}

\item{streak_x_param}{Streak parameters}

\item{streak_size}{Size of streaks}

\item{layer}{Layer assignments}

\item{layer_param}{Layer parameters}

\item{seed}{Random seed}
}
\value{
Vector of expression values
}
\description{
Generate expression pattern
}
