% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_spot_count}
\alias{generate_spot_count}
\title{Generate spot-level count matrix}
\usage{
generate_spot_count(true_count, nGenes, cell_number, celltype_proportion)
}
\arguments{
\item{true_count}{List of true count matrices}

\item{nGenes}{Number of genes}

\item{cell_number}{Number of cells per spot}

\item{celltype_proportion}{Cell type proportions}
}
\value{
Matrix of spot-level counts
}
\description{
Generate spot-level count matrix
}
