% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{generate_strap_layer}
\alias{generate_strap_layer}
\title{Generate a single layer stripe}
\usage{
generate_strap_layer(coord, x_start, x_end, y_start, y_end, layer_id)
}
\arguments{
\item{coord}{Coordinate matrix}

\item{x_start}{Start x position}

\item{x_end}{End x position}

\item{y_start}{Start y position}

\item{y_end}{End y position}

\item{layer_id}{Layer identifier}
}
\value{
Data frame with layer assignments
}
\description{
Generate a single layer stripe
}
\keyword{internal}
