% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{generate_z_matrix}
\alias{generate_z_matrix}
\title{Generate z-score matrix}
\usage{
generate_z_matrix(expr, mu_g, theta_g, block = 1024, n_cores = 1, eps = 1e-06)
}
\arguments{
\item{expr}{Expression matrix}

\item{mu_g}{Mean parameters}

\item{theta_g}{Dispersion parameters}

\item{block}{Block size for processing}

\item{n_cores}{Number of cores}

\item{eps}{Small value to prevent zeros}
}
\value{
Matrix of z-scores
}
\description{
Generate z-score matrix
}
\keyword{internal}
