% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{moran_flag}
\alias{moran_flag}
\title{Moran's I flag}
\usage{
moran_flag(y, listw, thr = 0.15)
}
\arguments{
\item{y}{Response variable}

\item{listw}{Spatial weights list}

\item{thr}{Threshold for flagging}
}
\value{
Logical vector indicating flagged genes
}
\description{
Moran's I flag
}
\keyword{internal}
