% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{prepare_mu_mat}
\alias{prepare_mu_mat}
\title{Prepare mu matrix}
\usage{
prepare_mu_mat(expr, mu_g, lib_factor)
}
\arguments{
\item{expr}{Expression matrix}

\item{mu_g}{Mean parameters}

\item{lib_factor}{Library size factors}
}
\value{
Matrix of mu values
}
\description{
Prepare mu matrix
}
\keyword{internal}
