% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{randcop}
\alias{randcop}
\title{Sample from a gaussian copula}
\usage{
randcop(cov_matrix, nSpots, epsilon = 1e-06)
}
\arguments{
\item{cov_matrix}{Correlation matrix for gaussian copula}

\item{nSpots}{Number of samples to draw}
}
\value{
Matrix of samples from gaussian copula
}
\description{
Sample from a gaussian copula
}
\keyword{internal}
