% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{refit_nb_gam_for_spatial_genes}
\alias{refit_nb_gam_for_spatial_genes}
\title{Refit NB-GAM for spatial genes}
\usage{
refit_nb_gam_for_spatial_genes(
  expr,
  coord,
  lib_factor,
  spatial_genes,
  mu_mat,
  theta_g,
  spline_k = 50
)
}
\arguments{
\item{expr}{Expression matrix}

\item{coord}{Coordinates data}

\item{lib_factor}{Library size factors}

\item{spatial_genes}{Spatial genes}

\item{mu_mat}{Mu matrix}

\item{theta_g}{Theta parameters}

\item{spline_k}{Basis dimension for the thin-plate spline.}
}
\value{
Updated mu_mat and theta_g
}
\description{
Refit NB-GAM for spatial genes
}
\keyword{internal}
