% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STCOS-package.R
\docType{package}
\name{STCOS-package}
\alias{STCOS}
\alias{STCOS-package}
\title{ST-COS: Spatial Transcriptomics Cell-type Oriented Simulator}
\description{
A simulator for spatial transcriptomics data with configurable spatial patterns, cell-type-specific expression profiles, and explicit latent Gaussian pairwise-dependence parameters. Realized count-scale dependence is generated through negative-binomial marginals and is not assumed to equal the latent target.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Yiyannnnn/ST-COS}
  \item Report bugs at \url{https://github.com/Yiyannnnn/ST-COS/issues}
}

}
\author{
\strong{Maintainer}: Yiyan Zhang \email{yiyan@unc.edu}

}
\keyword{internal}
