% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{stcos}
\alias{stcos}
\alias{stcos-package}
\title{stcos: Space-Time Change of Support}
\description{
An R Package for Space-Time Change of Support (STCOS) modeling.
}
\details{
Supports building and running STCOS and related models. A guide
on package use is given in <arXiv:1904.12092>.
}
\references{
Jonathan R. Bradley, Christopher K. Wikle, and Scott H. Holan (2015).
Spatio-temporal change of support with application to American Community
Survey multi-year period estimates. STAT 4 pp.255-270.
\doi{10.1002/sta4.94}.

Andrew M. Raim, Scott H. Holan, Jonathan R. Bradley, and Christopher K.
Wikle (2017). A model selection study for spatio-temporal change of
support. In JSM Proceedings, Government Statistics Section. Alexandria,
VA: American Statistical Association, pp.1524-1540.

Andrew M. Raim, Scott H. Holan, Jonathan R. Bradley, and Christopher K.
Wikle (2020+). Spatio-Temporal Change of Support Modeling with R.
\url{https://arxiv.org/abs/1904.12092}.
}
